+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.6.11] - 2021-03-19
 ### Added
 - A guide book... Don't get too excited, it only gives a link to the wiki right now
+### Changed
+- Model and texture debug logging (`debug.logging.modelAndTexture`) is now disabled by default, since no recent reports of the texture issue have been made. Update your config if desired.
 
 ## [2.6.10] - 2021-03-18
 ### Added